Minor Fixes
