+++ api/controllers/folderController.js
@@ -18,6 +18,7 @@ export const folderCreation =  async (req , res) =>{
         })
 
         const newFolder = await folder.save();
+        console.log(newFolder)
         
         const response = await User.findByIdAndUpdate(
             userObjId,
@@ -27,7 +28,7 @@ export const folderCreation =  async (req , res) =>{
                 }
             }
         )
-
+        console.log(response)  
         res.status(200).json(successMessage("folder created" , [newFolder,response]));
     }catch(error){
         console.log(`error: ${error}`);
