+++ client/src/components/Header.jsx
@@ -16,6 +16,11 @@ import { Avatar, Badge, Input, Dropdown, DropdownItem, WindmillContext } from '@
 function Header() {
   const { mode, toggleMode } = useContext(WindmillContext);
   const { toggleSidebar } = useContext(SidebarContext);
+
+  const [query,setquery]=useState("");
+
+  
+
   const navigate = useNavigate();
 
   const [isNotificationsMenuOpen, setIsNotificationsMenuOpen] = useState(false);
@@ -29,7 +34,9 @@ function Header() {
     setIsProfileMenuOpen(!isProfileMenuOpen);
   }
 
-  const [search,setSearch]=useState('');
+;
+
+  console.log("query",query)
 
   const logout = () => {
     navigate('/login');
@@ -50,9 +57,9 @@ function Header() {
         <div className="flex justify-center flex-1 lg:mr-32">
           <div className="relative w-full max-w-xl mr-6 focus-within:text-purple-500">
             <div className="absolute inset-y-0 flex items-center pl-2">
-              <SearchIcon className="w-4 h-4" aria-hidden="true" />
+              <SearchIcon className="w-4 h-4 b" aria-hidden="true"  />
             </div>
-            <Input className="pl-8 text-gray-700" placeholder="Search for documents" aria-label="Search" />
+            <Input className="pl-8 text-gray-700" placeholder="Search for documents"  onChange={(e)=>setquery(e.target.value)} aria-label="Search" />
           </div>
         </div>
         <ul className="flex items-center flex-shrink-0 space-x-6">
