--- client/src/components/Header1.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function Header() {
-  return (
-    <div>Header</div>
-  )
-}
