--- client/src/components/Login.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Login() {
-  return (
-    <div>Login</div>
-  )
-}
-
-export default Login
